[Refactoring] modify domain layer
