+++ app/src/main/java/com/example/sopt_seminar/data/repository/UserRepositoryImpl.kt
@@ -2,6 +2,7 @@ package com.example.sopt_seminar.data.repository
 
 import com.example.sopt_seminar.data.model.User
 import com.example.sopt_seminar.data.source.local.UserLocalDatSource
+import com.example.sopt_seminar.domain.repository.UserRepository
 import javax.inject.Inject
 import kotlinx.coroutines.flow.Flow
