+++ app/src/main/java/com/example/sopt_seminar/domain/repository/UserRepository.kt
@@ -1,4 +1,4 @@
-package com.example.sopt_seminar.data.repository
+package com.example.sopt_seminar.domain.repository
 
 import com.example.sopt_seminar.data.model.User
 import kotlinx.coroutines.flow.Flow
