+++ app/src/main/java/com/example/sopt_seminar/domain/usecase/SetUserUseCase.kt
@@ -1,7 +1,7 @@
 package com.example.sopt_seminar.domain.usecase
 
 import com.example.sopt_seminar.data.model.User
-import com.example.sopt_seminar.data.repository.UserRepository
+import com.example.sopt_seminar.domain.repository.UserRepository
 import javax.inject.Inject
 
 class SetUserUseCase @Inject constructor(
