+++ app/src/main/java/com/example/sopt_seminar/util/Event.kt
@@ -1,4 +1,4 @@
-package com.example.sopt_seminar.domain.state
+package com.example.sopt_seminar.util
 
 /**
  * Used as a wrapper for data that is exposed via a LiveData that represents an event.
